+++ order/build.gradle
@@ -13,6 +13,9 @@ dependencies {
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    
+    // Apache HttpClient 5 for RestClient HttpComponentsClientHttpRequestFactory
+    implementation 'org.apache.httpcomponents.client5:httpclient5'
 
     implementation 'io.micrometer:micrometer-tracing-bridge-brave'
     implementation 'io.zipkin.reporter2:zipkin-reporter-brave'
