+++ payment/src/main/java/com/ll/payment/payment/service/PaymentService.java
@@ -13,9 +13,4 @@ public interface PaymentService {
 
     Payment refundPayment(PaymentRefundRequest request);
 
-    Payment completeDepositPayment(PaymentRequest payment, int amount);
-
-    Payment chargeDepositWithToss(PaymentRequest payment, int chargeAmount);
-
-    Payment processDepositPaymentWithCharge(PaymentRequest payment, int shortageAmount, int requestedAmount);
 }
